Move ColumnKeyValue to the data loader core project (#1770)
