+++ data-loader/cli/src/main/java/com/scalar/db/dataloader/cli/command/ColumnKeyValueConverter.java
@@ -1,5 +1,6 @@
 package com.scalar.db.dataloader.cli.command;
 
+import com.scalar.db.dataloader.core.ColumnKeyValue;
 import picocli.CommandLine;
 
 /**
